backend/requirements.txt
fastapi==0.103.1
uvicorn[standard]==0.23.2
SQLAlchemy==2.0.21
psycopg2-binary==2.9.7
python-dotenv==1.0.0
